+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## latest
 
+* [2017-12-15] Minimum parameter values can now be set and verified with
+  `utils.set_minimum` and `utils.get_minimum`.
+
 * [2017-11-28] New Hankel filter `wer_201_2018`
 
 * [2017-11-09] `opt=parallel` has no effect if `numexpr` is not built against